splitted Phase2TrackerFEDBuffer.h as requested + cleaning
