+++ EventFilter/Phase2TrackerRawToDigi/BuildFile.xml
@@ -9,7 +9,6 @@
 <use   name="FWCore/ParameterSet"/>
 <use   name="FWCore/PluginManager"/>
 <use   name="FWCore/Utilities"/>
-<use   name="boost"/>
 <export>
   <lib   name="1"/>
 </export>
