+++ EventFilter/Phase2TrackerRawToDigi/plugins/BuildFile.xml
@@ -13,6 +13,5 @@
   <use   name="FWCore/Framework"/>
   <use   name="FWCore/MessageLogger"/>
   <use   name="FWCore/ParameterSet"/>
-  <use   name="boost"/>
   <flags   EDM_PLUGIN="1"/>
 </library>
