+++ EventFilter/Phase2TrackerRawToDigi/plugins/Phase2TrackerDigiProducer.h
@@ -12,10 +12,9 @@
 #include "DataFormats/Common/interface/DetSetVector.h"
 #include "DataFormats/DetId/interface/DetIdCollection.h"
 #include "DataFormats/FEDRawData/interface/FEDRawDataCollection.h"
-#include "EventFilter/Phase2TrackerRawToDigi/interface/Phase2TrackerFEDBuffer.h"
 #include "CondFormats/SiStripObjects/interface/Phase2TrackerCabling.h"
 #include "DataFormats/Phase2TrackerDigi/interface/Phase2TrackerDigi.h"
-#include "boost/cstdint.hpp"
+#include <stdint.h>
 #include <iostream>
 #include <string>
 #include <vector>
@@ -34,14 +33,25 @@ namespace Phase2Tracker {
     virtual void produce(edm::Event&, const edm::EventSetup&) override;
     virtual void endJob() override;
 
-    
   private:
     unsigned int runNumber_;
     edm::EDGetTokenT<FEDRawDataCollection> token_;
     const Phase2TrackerCabling * cabling_;
     uint32_t cacheId_;
     DetIdCollection detids_;
-
+    class Registry {
+      public:
+        /// constructor
+        Registry(uint32_t aDetid, uint16_t firstStrip, size_t indexInVector, uint16_t numberOfDigis) :
+          detid(aDetid), first(firstStrip), index(indexInVector), length(numberOfDigis) {}
+        /// < operator to sort registries
+        bool operator<(const Registry &other) const {return (detid != other.detid ? detid < other.detid : first < other.first);}
+        /// public data members
+        uint32_t detid;
+        uint16_t first;
+        size_t index;
+        uint16_t length;
+    };
     std::vector<Registry> proc_work_registry_;
     std::vector<Phase2TrackerDigi> proc_work_digis_;
   };
